requirements.txt
Django
psycopg2-binary>=2.8
djangorestframework
django-filter